Float to Double.

+++ GeneratorInterface/PyquenInterface/interface/PyquenWrapper.h
@@ -4,6 +4,7 @@
 /*
  *
  * Wrapper for FORTRAN version of PYQUEN
+ * $Id:$
  *
  * Camelia Mironov
  *
@@ -12,22 +13,21 @@
 // PYQUEN routine declaration
 
 extern "C" {
-  void pyquen_(float& a,int& ifb,float& bfix);
+  void pyquen_(double& a,int& ifb,double& bfix);
 }
 #define PYQUEN pyquen_ 
 
 extern "C" {
   extern struct{
-    float bgen;
+    double bgen;
   }plfpar_;
 }
 #define plfpar plfpar_
 
-
 extern "C" {
   extern struct{
-    float T0u;
-    float tau0u;
+    double T0u;
+    double tau0u;
     int   nfu;
     int   ienglu;
     int   ianglu;
